solve: MST 두문제 풀이 -프림 방식, 크루스칼 방식 -각 방식에 맞는 문제 풀이용 구조체 코드 정리
